[TS-286] Refactor: star 관련 상수 파일을 starPageConstants.ts로 합침
상수가 얼마 되지 않아서 찾기 편하게 한 파일로 합쳤음
