--- src/constants/starCardDetailConstants.ts
@@ -1,5 +1,0 @@
-export const buttonState = {
-	default: "선택",
-	selected: "별자리 보러 가기",
-	have: "프로필 캐릭터로 설정",
-};
