+++ src/constants/starPageConstants.ts
@@ -1,3 +1,9 @@
+export const buttonState = {
+	default: "선택",
+	selected: "별자리 보러 가기",
+	have: "프로필 캐릭터로 설정",
+};
+
 export const categoryData = [
 	"전체",
 	"육체 활동",
